// Inventroy/inventory/src/app/common/models/common-data-list.model.ts
export class CommonDataList {
    lstCountries: any;
    lstDivisions: any;
    lstDistricts: any;
    lstUpazilas: any;
    lstPostOffices: any;
    lstRelationShips: any;
    lstNationalities: any;
    lstProfessions: any;
    lstShops: any;
    lstOrganizations: any;
    lstShiftSections: any;
    lstEmployees: any;
    lstCategories:any;
    lstProducts:any;
    lstVats:any;
    lstUnits:any;
    lstManufacturers:any;
    lstSpecificationAttributes:any;
    lstSpecificationAttrValues:any;
}